functions/msg-to-html-body/requirements.unexpanded.txt
dataclasses==0.6
gobits==1.0.7
google-cloud-pubsub==2.3.0
Jinja2==2.11.3